pyproject.toml
[project]
name = "hamster-wheel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-escpos[all]>=3.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "tomli>=2.0.1; python_version < '3.11'",
    "python-multipart>=0.0.20",
    "Pillow>=10.0.0",
    "schedule>=1.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
